refactor(repository): split card.go by concern (same package)
